requirements.txt
django
djangorestframework 
django-cors-headers 
django-rest-swagger
jwt 
drf_yasg
